+++ app/src/main/java/ba/unsa/etf/rma/rma20djokicmilica36/TransactionListAdapter.java
@@ -49,7 +49,7 @@ public class TransactionListAdapter extends ArrayAdapter<Transaction> {
        iznos = (TextView) newView.findViewById(R.id.iznos);
 
        title.setText(transaction.getTitle());
-       iznos.setText(Integer.toString(transaction.getAmount()));
+       iznos.setText(Double.toString(transaction.getAmount()));
 
        transactionType tip = transaction.getType();
